Extract chat onboarding helpers
