+++ tests/test-correctness-phase2.js
@@ -131,6 +131,7 @@ const pwaAppShellAssets = [
   '/js/chat-actions.js',
   '/js/chat-panel.js',
   '/js/chat-discussion.js',
+  '/js/chat-onboarding.js',
   '/js/chat-icons.js',
   '/js/chat-personalities.js',
   '/js/chat-history.js',
